+++ src/Component.hpp
@@ -106,7 +106,7 @@ public:
 	// @param list list to populate
 	template <typename T>
 	void findAllComponents(Component::type_t type, LinkedList<T*>& list) const {
-		for( size_t c = 0; c < components.getSize(); ++c ) {
+		for( Uint32 c = 0; c < components.getSize(); ++c ) {
 			if( components[c]->getType() == type ) {
 				list.addNodeLast(dynamic_cast<T*>(components[c]));
 			}
@@ -122,7 +122,7 @@ public:
 		if( name == nullptr || strcmp(name,"")==0 ) {
 			return nullptr;
 		}
-		for( size_t c = 0; c < components.getSize(); ++c ) {
+		for( Uint32 c = 0; c < components.getSize(); ++c ) {
 			if( strcmp( components[c]->getName(), name ) == 0 ) {
 				return static_cast<T*>(components[c]);
 			} else {
@@ -143,7 +143,7 @@ public:
 		if( uid == nuid ) {
 			return nullptr;
 		}
-		for( size_t c = 0; c < components.getSize(); ++c ) {
+		for( Uint32 c = 0; c < components.getSize(); ++c ) {
 			if( components[c]->getUID() == uid ) {
 				return static_cast<T*>(components[c]);
 			} else {
